+++ packages/core/src/diff/rules/index.ts
@@ -1,2 +1,3 @@
+export * from './dangerous-breaking';
 export * from './suppress-removal-of-deprecated-field';
 export * from './ignore-description-changes';